% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{nonParametric}
\alias{nonParametric}
\title{Perform non-parametric test}
\usage{
nonParametric(data)
}
\arguments{
\item{data}{input the dataset that was used for clustering along with cluster results in column called cluster}
}
\value{
returns a list with non-parametric test results and p-values
}
\description{
This function performs non-parametric test on clustered data. Name your cluster column as 'cluster'.
}
\examples{
# this code will run

# data(iris)
# data = iris[,1:4]

# cluster = kmeans(data,4)$cluster

# nonParametric(data)
}
