% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{sampleSize}
\alias{sampleSize}
\title{create sample sizes}
\usage{
sampleSize(data, cluster)
}
\arguments{
\item{data}{input the dataset that was used for clustering}

\item{cluster}{enter the cluster column of the data}
}
\value{
returns a dataframe with equal samples
}
\description{
Creates a sample size of 30 for each group. If in case the population group sample is less than 30 then, the sample size will be the minimum of group sample size.
}
\examples{
# this code will run

# data(iris)
# data = iris[,1:4]

# cluster = kmeans(data,4)$cluster

# sampleSize(data, cluster)
}
