% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNormality.R
\name{TestNormality}
\alias{TestNormality}
\title{Test normality}
\usage{
TestNormality(data)
}
\arguments{
\item{data}{input the dataset that was used for clustering}
}
\value{
returns a list with normality test results and p-values
}
\description{
This function performs Shapiro-Wilk test is used to test the normality of the sample population.

Null Hypothesis H0 for this test is: the distribution of the data are not significantly different from normal distribution. In other words, we can assume the normality
}
\examples{
# this code will run

# data(iris)
# data = iris[,1:4]

# cluster = kmeans(data,4)$cluster

# TestNormality(data)
}
